ml_collector/requirements.txt
# Minimal dependencies for ML data collector
websockets==12.0
python-dotenv==1.0.0
